// FrontFredy/src/app/model/quotes-bookings-report.ts
export class QuotesBookingsReport {
    id: Number;
    type: String;
    requirementDate: Date;
    assignmentDate: Date;
    answerDate: Date;
    user: string;
    company: String;
    userName: String;
    origin: String;
    destination: String;
    customerType: String;
    quantityPieces: Number;
    commodity: String;
    stackable: String;
    chargableweight: Number;
    rate: Number;
    status: String;
    temperatureControl: String;
}